TIP-842: Move association type integration test to structure sub context
