+++ src/Akeneo/Pim/Structure/Component/tests/Integration/Normalizer/Standard/AssociationTypeIntegration.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace tests\integration\Pim\Component\Catalog\Normalizer\Standard;
+namespace Akeneo\Pim\Structure\Component\tests\Integration\Normalizer\Standard;
 
-use Pim\Component\Catalog\tests\integration\Normalizer\Standard\AbstractStandardNormalizerTestCase;
+use Akeneo\Test\Integration\TestCase;
 
 /**
  * @author    Marie Bochu <marie.bochu@akeneo.com>
  * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class AssociationTypeIntegration extends AbstractStandardNormalizerTestCase
+class AssociationTypeIntegration extends TestCase
 {
     public function testAssociationType()
     {
@@ -27,5 +27,13 @@ class AssociationTypeIntegration extends AbstractStandardNormalizerTestCase
         $result = $serializer->normalize($repository->findOneByIdentifier('SUBSTITUTION'), 'standard');
 
         $this->assertSame($expected, $result);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function getConfiguration()
+    {
+        return $this->catalog->useTechnicalCatalog();
     }
 }
